Semantic time values instead of mid-dates
Represent month/quarter/year as that thing, not its mid-date. This is
more appropriate default behavior, especially for R users. The downside
is that year-week values have no specific representation in zoo, so we
represent them with as opaque strings with correct ordering, so that the
X axis labeling in default plots for weekly datasets output by
`dmseries` is not pretty like it is for the other time granularities.